payment/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
Werkzeug<2.1.0
mysql-connector-python==8.0.26